+++ Kernel/Output/HTML/Dashboard/UserOffline.pm
@@ -216,7 +216,6 @@ sub Run {
 
             # get session data
             my %Data = $SessionObject->GetSessionIDData( SessionID => $SessionID );
-
             next SESSIONID if !%Data;
             next SESSIONID if !$Data{UserID};
 
@@ -237,12 +236,6 @@ sub Run {
                     UserLogin => $Data{UserLogin},
                     );
             }
-
-            # only show if not already shown
-            # KIX4OTRS-capeIT
-            # next SESSIONID if $Online->{User}->{ $Data{UserType} }->{ $Data{UserID} };
-            next SESSIONID if $Offline->{User}->{ $Data{UserType} }->{ $Data{UserID} };
-            # EO KIX4OTRS-capeIT
 
             # check last request time / idle time out
             next SESSIONID if !$Data{UserLastRequest};
@@ -354,51 +347,6 @@ sub Run {
     # get config object
     my $ConfigObject = $Kernel::OM->Get('Kernel::Config');
 
-    # Check if agent has permission to start chats with the listed users
-    my $EnableChat               = 1;
-    my $ChatStartingAgentsGroup  = $ConfigObject->Get('ChatEngine::PermissionGroup::ChatStartingAgents') || 'users';
-    my $ChatReceivingAgentsGroup = $ConfigObject->Get('ChatEngine::PermissionGroup::ChatReceivingAgents') || 'users';
-
-    if (
-        !$ConfigObject->Get('ChatEngine::Active')
-        || !defined $LayoutObject->{"UserIsGroup[$ChatStartingAgentsGroup]"}
-        || $LayoutObject->{"UserIsGroup[$ChatStartingAgentsGroup]"} ne 'Yes'
-        )
-    {
-        $EnableChat = 0;
-    }
-    if (
-        $EnableChat
-        && $Self->{Filter} eq 'Agent'
-        && !$ConfigObject->Get('ChatEngine::ChatDirection::AgentToAgent')
-        )
-    {
-        $EnableChat = 0;
-    }
-    if (
-        $EnableChat
-        && $Self->{Filter} eq 'Customer'
-        && !$ConfigObject->Get('ChatEngine::ChatDirection::AgentToCustomer')
-        )
-    {
-        $EnableChat = 0;
-    }
-
-    my $VideoChatEnabled = 0;
-    my $VideoChatAgentsGroup = $ConfigObject->Get('ChatEngine::PermissionGroup::VideoChatAgents') || 'users';
-
-    # Enable the video chat feature if system is entitled and agent is a member of configured group.
-    if (
-        $ConfigObject->Get('ChatEngine::Active')
-        && defined $LayoutObject->{"UserIsGroup[$VideoChatAgentsGroup]"}
-        && $LayoutObject->{"UserIsGroup[$VideoChatAgentsGroup]"} eq 'Yes'
-        )
-    {
-        if ( $Kernel::OM->Get('Kernel::System::Main')->Require( 'Kernel::System::VideoChat', Silent => 1 ) ) {
-            $VideoChatEnabled = $Kernel::OM->Get('Kernel::System::VideoChat')->IsEnabled();
-        }
-    }
-
     USERID:
 
     # KIX4OTRS-capeIT
@@ -417,11 +365,6 @@ sub Run {
         # my $UserData = $OnlineData{$UserID};
         my $UserData = $OfflineData{$UserID};
         # EO KIX4OTRS-capeIT
-        my $AgentEnableChat    = 0;
-        my $CustomerEnableChat = 0;
-        my $ChatAccess         = 0;
-        my $VideoChatAvailable = 0;
-        my $VideoChatSupport   = 0;
 
         # Default status
         # KIX4OTRS-capeIT
@@ -430,72 +373,6 @@ sub Run {
         # EO KIX4OTRS-capeIT
         my $UserStateDescription = $LayoutObject->{LanguageObject}->Translate('This user is currently offline');
 
-        # we also need to check if the receiving agent has chat permissions
-        if ( $EnableChat && $Self->{Filter} eq 'Agent' ) {
-
-            my %UserGroups = $Kernel::OM->Get('Kernel::System::Group')->PermissionUserGet(
-                UserID => $UserData->{UserID},
-                Type   => 'rw',
-            );
-
-            my %UserGroupsReverse = reverse %UserGroups;
-            $ChatAccess = $UserGroupsReverse{$ChatReceivingAgentsGroup} ? 1 : 0;
-
-            my %User = $Kernel::OM->Get('Kernel::System::User')->GetUserData(
-                UserID => $UserID,
-            );
-            $VideoChatSupport = $User{VideoChatHasWebRTC};
-
-            # Check agents availability
-            if ($ChatAccess) {
-                my $AgentChatAvailability = $Kernel::OM->Get('Kernel::System::Chat')->AgentAvailabilityGet(
-                    UserID   => $UserID,
-                    External => 0,
-                );
-
-                if ( $AgentChatAvailability == 3 ) {
-                    $UserState            = Translatable('Active');
-                    $AgentEnableChat      = 1;
-                    $UserStateDescription = $LayoutObject->{LanguageObject}->Translate('User is currently active.');
-                    $VideoChatAvailable   = 1;
-                }
-                elsif ( $AgentChatAvailability == 2 ) {
-                    $UserState       = Translatable('Away');
-                    $AgentEnableChat = 1;
-                    $UserStateDescription
-                        = $LayoutObject->{LanguageObject}->Translate('User was inactive for a while.');
-                }
-                elsif ( $AgentChatAvailability == 1 ) {
-                    $UserState = Translatable('Unavailable');
-                    $UserStateDescription
-                        = $LayoutObject->{LanguageObject}->Translate('User set their status to unavailable.');
-                }
-            }
-        }
-        elsif ($EnableChat) {
-            $ChatAccess = 1;
-
-            my $CustomerChatAvailability = $Kernel::OM->Get('Kernel::System::Chat')->CustomerAvailabilityGet(
-                UserID => $UserData->{UserID},
-            );
-
-            my %CustomerUser = $Kernel::OM->Get('Kernel::System::CustomerUser')->CustomerUserDataGet(
-                User => $UserID,
-            );
-            $VideoChatSupport = 1 if $CustomerUser{VideoChatHasWebRTC};
-
-            if ( $CustomerChatAvailability == 3 ) {
-                $UserState            = Translatable('Active');
-                $CustomerEnableChat   = 1;
-                $UserStateDescription = $LayoutObject->{LanguageObject}->Translate('User is currently active.');
-                $VideoChatAvailable   = 1;
-            }
-            elsif ( $CustomerChatAvailability == 2 ) {
-                $UserState            = Translatable('Away');
-                $CustomerEnableChat   = 1;
-                $UserStateDescription = $LayoutObject->{LanguageObject}->Translate('User was inactive for a while.');
-            }
-        }
         $LayoutObject->Block(
 
             # KIX4OTRS-capeIT
@@ -505,14 +382,8 @@ sub Run {
             # EO KIX4OTRS-capeIT
             Data => {
                 %{$UserData},
-                ChatAccess           => $ChatAccess,
-                AgentEnableChat      => $AgentEnableChat,
-                CustomerEnableChat   => $CustomerEnableChat,
                 UserState            => $UserState,
                 UserStateDescription => $UserStateDescription,
-                VideoChatEnabled     => $VideoChatEnabled,
-                VideoChatAvailable   => $VideoChatAvailable,
-                VideoChatSupport     => $VideoChatSupport,
             },
         );